Refactor: centralize schema handling logic in `SchemaUtil` and clean up `MultipartFormDataValidator`
